move home to pbr, add proj-res filter
