+++ lib/platform-bible-react/src/preview/pages/components/basics/projectResourceFilter.examples.component.tsx
@@ -1,0 +1,10 @@
+import ProjectResourceFilter from '@/components/basics/projectResourceFilter.component';
+
+export function ProjectResourceFilterExamples() {
+  return (
+    <>
+      <ProjectResourceFilter onChange={() => {}} />
+      <ProjectResourceFilter onChange={() => {}} variant="ghost" />
+    </>
+  );
+}
